requirements.txt
requests
google-genai
python-dotenv
uvicorn
fastapi
pydantic
google-api-python-client
google-auth-httplib2
google-auth-oauthlib